Finish styling CollectibleDetailsDrawer
